large size update

+++ database/migrations/2022_07_19_175945_create_folder_user_table.php
@@ -13,15 +13,18 @@ class CreateFolderUserTable extends Migration
      */
     public function up()
     {
-        Schema::create('folder_user', function (Blueprint $table) {
-            $table->unsignedBigInteger('user_id');
-            $table->unsignedBigInteger('folder_id');
-        });
 
-        Schema::table('folder_user', function($table) {
-            $table->foreign('user_id', 'user_id_fk_6885')->references('id')->on('users')->onDelete('cascade');
-            $table->foreign('folder_id', 'folder_id_fk_6885')->references('id')->on('folders')->onDelete('cascade');
-        });
+        if (!Schema::hasTable('folder_user')) {
+            Schema::create('folder_user', function (Blueprint $table) {
+                $table->unsignedBigInteger('user_id');
+                $table->unsignedBigInteger('folder_id');
+            });
+
+            Schema::table('folder_user', function($table) {
+                $table->foreign('user_id', 'user_id_fk_6885')->references('id')->on('users')->onDelete('cascade');
+                $table->foreign('folder_id', 'folder_id_fk_6885')->references('id')->on('folders')->onDelete('cascade');
+            });
+        };
 
     }
